perf(transformRequest): fast-path watch and sourcemap handling (#16170)
